pyproject.toml
[tool.poetry]
name = "demo-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = "^2.0.31"
pymysql = "^1.1.1"
alembic = "^1.13.3"
aiomysql = "^0.2.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.12"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
langserve = {extras = ["all"], version = "^0.3.0"}
langchain-openai = "^0.2.5"
langchain = "^0.3.7"
google = "^3.0.0"
google-auth = "^2.36.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
